+++ src/app/modules/market/component/market-item-search-filter-heist/market-item-search-filter-heist.component.ts
@@ -1,0 +1,12 @@
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { TradeSearchRequest } from '@shared/module/poe/trade';
+
+@Component({
+  selector: 'app-market-item-search-filter-heist',
+  templateUrl: './market-item-search-filter-heist.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class MarketItemSearchFilterHeistComponent {
+  @Input()
+  public request: TradeSearchRequest;
+}