+++ src/app/modules/replay/component/replay-settings/replay-settings.component.ts
@@ -1,6 +1,8 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { FeatureSettingsComponent } from '@app/feature';
+import { OWReplay } from '@app/odk';
 import { ReplayFeatureSettings } from '@modules/replay/replay-feature-settings';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-replay-settings',
@@ -8,8 +10,14 @@ import { ReplayFeatureSettings } from '@modules/replay/replay-feature-settings';
   styleUrls: ['./replay-settings.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ReplaySettingsComponent extends FeatureSettingsComponent<ReplayFeatureSettings> {
+export class ReplaySettingsComponent extends FeatureSettingsComponent<ReplayFeatureSettings> implements OnInit {
+  public path$: Observable<string>;
+
   public displayDuration = (value: number) => `${value}s`;
+
+  public ngOnInit(): void {
+    this.path$ = OWReplay.path();
+  }
 
   public load(): void { }